core: fix C++ compatibility warnings
Use an int as the last named libusb_set_option parameter so va_start does
not receive a type subject to default argument promotion. Validate both
option bounds before converting back to the backend enum type.

Cast the Apple request-type inputs to their byte representation before the
SDK macro combines values from unrelated anonymous enums. Add coverage for
invalid option bounds.

The C++ build added by #1845 exposed these issues.

+++ libusb/os/darwin_usb.c
@@ -1130,7 +1130,8 @@ static int darwin_fetch_string_descriptor (struct libusb_context *ctx,
   langid = dpriv->langid;
   if (0 == langid) {
     memset (desc, 0, sizeof(desc));
-    req.bmRequestType = USBmakebmRequestType (kUSBIn, kUSBStandard, kUSBDevice);
+    req.bmRequestType = USBmakebmRequestType ((UInt8)kUSBIn,
+      (UInt8)kUSBStandard, (UInt8)kUSBDevice);
     req.bRequest      = kUSBRqGetDescriptor;
     req.wValue        = (UInt16)(LIBUSB_DT_STRING << 8);
     req.wIndex        = 0;
@@ -1147,7 +1148,8 @@ static int darwin_fetch_string_descriptor (struct libusb_context *ctx,
 
   /* Fetch the requested string descriptor. */
   memset (desc, 0, sizeof(desc));
-  req.bmRequestType = USBmakebmRequestType (kUSBIn, kUSBStandard, kUSBDevice);
+  req.bmRequestType = USBmakebmRequestType ((UInt8)kUSBIn,
+    (UInt8)kUSBStandard, (UInt8)kUSBDevice);
   req.bRequest      = kUSBRqGetDescriptor;
   req.wValue        = (UInt16)((LIBUSB_DT_STRING << 8) | string_index);
   req.wIndex        = langid;
@@ -1369,7 +1371,8 @@ static IOReturn darwin_request_descriptor (usb_device_t device, UInt8 desc, UInt
   memset (buffer, 0, buffer_size);
 
   /* Set up request for descriptor/ */
-  req.bmRequestType = USBmakebmRequestType(kUSBIn, kUSBStandard, kUSBDevice);
+  req.bmRequestType = USBmakebmRequestType((UInt8)kUSBIn,
+    (UInt8)kUSBStandard, (UInt8)kUSBDevice);
   req.bRequest      = kUSBRqGetDescriptor;
   req.wValue        = (UInt16)(desc << 8);
   req.wIndex        = desc_index;
